Advanced Graph Functions

Some more advanced functions for quering a graph.

Imported to namespace by

using NaiveNASlib.Advanced

NaiveNASlib.findterminating — Function

findterminating(v::AbstractVertex, direction::Function, other::Function= v -> [], visited = [])

Return an array of all vertices which terminate size changes (i.e does not propagate them) seen through the given direction (typically inputs or outputs). A vertex will be present once for each unique path through which its seen.

The other direction may be specified and will be traversed if a SizeInvariant vertex is encountered.

Will return the given vertex if it is terminating.

Examples

julia> v1 = inputvertex("v1", 3);

julia> v2 = inputvertex("v2", 3);

julia> v3 = conc(v1,v2,v1,dims=1);

julia> name.(findterminating(v1, outputs, inputs))
1-element Array{String,1}:
 "v1"

julia> name.(findterminating(v3, outputs, inputs))
0-element Array{Any,1}

julia> name.(findterminating(v3, inputs, outputs))
3-element Array{String,1}:
 "v1"
 "v2"
 "v1"

 julia> v5 = v3 + inputvertex("v4", 9);

 julia> name.(findterminating(v3, outputs, inputs))
 1-element Array{String,1}:
  "v4"

NaiveNASlib.all_in_graph — Function

all_in_graph(v::AbstractVertex)

Return an array of vertices in the same graph (or connected component) as v

NaiveNASlib.output! — Function

output!(memo::Dict{AbstractVertex, Any}, v::AbstractVertex)

Return the output from v given any input in memo by traversing the graph. Intermediate results from all visited vertices will be stored in memo after function exits.

Examples

julia> using NaiveNASlib

julia> ivs = InputVertex.(1:2);

julia> cv = CompVertex(*, ivs...);

julia> results = Dict{AbstractVertex, Any}(zip(ivs, [2,3]));

julia> output!(results, CompVertex(-, cv, ivs[1]))
4
julia> results
Dict{AbstractVertex,Any} with 4 entries:
  CompVertex(*, [InputVertex(1), InputVertex(2)], [CompVertex(-)]) => 6
  CompVertex(-, [CompVertex(*), InputVertex(1)], [])               => 4
  InputVertex(1, [CompVertex(*), CompVertex(-)])                   => 2
  InputVertex(2, [CompVertex(*)])                                  => 3

NaiveNASlib.ancestors — Function

ancestors(v::AbstractVertex)

Return an array of all ancestors of v, including v itself.

Examples

```julia-repl julia> ancestors(CompVertex(+, InputVertex.(1:2)...)) 3-element Array{AbstractVertex,1}: InputVertex(1) InputVertex(2) CompVertex(+, [InputVertex(1), InputVertex(2)])

NaiveNASlib.descendants — Function

descendants(v::AbstractVertex)

Return an array of all descendants of v, including v itself.

Examples

```julia-repl julia> descendants(CompVertex(+, InputVertex.(1:2)...)) 3-element Array{AbstractVertex,1}: InputVertex(1) InputVertex(2) CompVertex(+, [InputVertex(1), InputVertex(2)])

NaiveNASlib.defaultutility — Function

defaultutility(v::AbstractVertex)

Default function used to calculate utility of output neurons when NeuronIndices is used and no function is provided.

Implement defaultutility(f) where f is the computation performed by CompVertex to set the default for f.

Examples

julia> struct Affine{T}
    W::Matrix{T}
end;

# Default is weight magnitude
julia> NaiveNASlib.defaultutility(l::Affine) = mean(abs, l.W; dims=2);