Element Wise Configuration

These are only to be used before >> when creating element wise vertices through infix operators (e.g. +, * etc.).

Imported to namespace by

using NaiveNASlib.Advanced

NaiveNASlib.VertexConf — Type

VertexConf
VertexConf(; traitdecoration = identity, outwrap = identity)

Config struct to be used with element wise op syntax (+, -, *, /).

traitdecoration allows for decorating the vertex trait with stuff like logging, validation etc. outwrap is a function which returns a function which will be applied to the computed output. For example, the following outwrap scales output by a factor of 2: outwrap = f -> (x...) -> 2f((x...)

NaiveNASlib.traitconf — Function

traitconf(t)

Shortcut for VertexConf(;traitdecoration=t).

NaiveNASlib.outwrapconf — Function

outwrapconf(o)

Shortcut for VertexConf(;outwrap=o).