Other Functions

Imported to namespace by

using NaiveNASlib.Extend

NaiveNASlib.parselect — Function

parselect(pars::AbstractArray{T,N}, elements_per_dim...; newfun = zeros) where {T, N}

Return a new AbstractArray{T, N} which has a subset of the elements of pars.

Which elements to select is determined by elements_per_dim which is a Pair{Int, Vector{Int}} mapping dimension (first memeber) to which elements to select in that dimension (second memeber).

For a single dim=>elems pair, the following holds: selectdim(output, dim, i) == selectdim(pars, dim, elems[i]) if elems[i] is positive and selectdim(output, dim, i) .== newfun(T, dim, size)[j] if elems[i] is the j:th negative value and size is sum(elems .< 0).

Examples

julia> pars = reshape(1:3*5, 3,5)
3×5 reshape(::UnitRange{Int64}, 3, 5) with eltype Int64:
 1  4  7  10  13
 2  5  8  11  14
 3  6  9  12  15

 julia> NaiveNASlib.parselect(pars, 1 => [-1, 1,3,-1,2], 2=>[3, -1, 2], newfun = (T, d, s...) -> -ones(T, s))
 5×3 Array{Int64,2}:
  -1  -1  -1
   7  -1   4
   9  -1   6
  -1  -1  -1
   8  -1   5

NaiveNASlib.vertex — Function

vertex(computation, trait::MutationTrait, inputs::AbstractVertex...; mutation=IoChange)

Return a mutable computation type vertex.

Examples

julia> using NaiveNASlib

julia> v = vertex(NaiveNASlib.SizeInvariant(), x -> 5x, inputvertex("input", 1));

julia> v(3)
15

NaiveNASlib.compconstraint! — Function

compconstraint!(case, s, v, data)

Add constraints on the computation (e.g. neural network layer) for AbstractVertex v using strategy s.

Extra info like the model and variables is provided in data.