Strategy Base Types

Useful when implementing new strategies.

Imported to namespace by

using NaiveNASlib.Extend

NaiveNASlib.AbstractΔSizeStrategy — Type

AbstractΔSizeStrategy

Abstract base type for strategies for how to change the size.

See Δsize!.

NaiveNASlib.AbstractJuMPΔSizeStrategy — Type

AbstractJuMPΔSizeStrategy <: AbstractΔSizeStrategy

Abstract type for strategies to change or align the sizes of vertices using JuMP.

NaiveNASlib.DecoratingJuMPΔSizeStrategy — Type

DecoratingJuMPΔSizeStrategy <: AbstractJuMPΔSizeStrategy

Abstract type for AbstractJuMPΔSizeStrategies which wants to delegate some parts of the problem formulation to another strategy.

More concretely: If s is a DecoratingJuMPΔSizeStrategy then base(s) will be used unless explicitly stated through dispatch.

NaiveNASlib.base — Method

base(s::DecoratingJuMPΔSizeStrategy)

Return the strategy wrapped by s.

NaiveNASlib.AbstractAfterΔSizeStrategy — Type

AbstractAfterΔSizeStrategy <: DecoratingJuMPΔSizeStrategy

Abstract base type for strategies which perform actions after size has changed (e.g validation and logging).

NaiveNASlib.AbstractAlignSizeStrategy — Type

AbstractAlignSizeStrategy

Base type for strategies for how to align size (nin/nout) when doing structural mutation.

Note that all strategies are not guaranteed to work in all cases.

Default strategies should however be selected based on case so that things always work out.

NaiveNASlib.AbstractConnectStrategy — Type

AbstractConnectStrategy

Base type for strategies for how to (re)connect vertices when doing structural mutation.