Internal functions with doc strings

These are not part of the public API.

NaiveNASlib.JuMPNorm — Type

JuMPNorm

Abstract type for norms to a JuMP model.

NaiveNASlib.sizeobjective! — Function

sizeobjective!(case, s::AbstractJuMPΔSizeStrategy, model, noutvars, sizetargets)

Add the objective for noutvars using strategy s.

NaiveNASlib.solve_outputs_selection — Function

solve_outputs_selection(s::AbstractΔSizeStrategy, vertices::AbstractVector{<:AbstractVertex}, utilityfun)

Returns a tuple (success, nindict, noutdict) where nindict[vi] are new input neuron indices and noutdict[vi] are new output neuron indices for each vertex vi in vertices.

The function generally tries to maximize sum(utilityfun(vi) .* selected[vi]) ∀ vi in vertices where selected[vi] is all elements in noutdict[vi] larger than 0 (negative values in noutdict indicates a new output shall be inserted at that position). This however is up to the implementation of the AbstractΔSizeStrategy s.

Since selection of outputs is not guaranteed to work in all cases, a flag success is also returned. If success is false then applying the new indices may (and probably will) fail.

NaiveNASlib.all_in_Δsize_graph — Function

all_in_Δsize_graph(v::AbstractVertex, d::Direction)

Return an array of vertices which will be affected if v changes size in direction d.

NaiveNASlib.noinsertgaps! — Function

noinsertgaps!(model, select, insert, maxinsert=length(outsel) * 10)

Add constraints so that insert does not create undefined gaps in the result of the neuron selection.

Assume select is a set of binary variables where select[i] = 1 means select the output neuron at position i and insert[i] = N means insert N new output neurons at the position after i.

An example of an undefined gap is if select = [1, 1, 0] and insert = [0, 0, 1] because this results in the instruction to use existing output neurons 1 and 2 and then insert a new neuron at position 4. In this example position 3 is an undefined gap as one should neither put an existing neuron there nor shall one insert new neurons. Running this method constrains model so that this solution is infeasible.

NaiveNASlib.Both — Type

Both

Represents both directions (Input and Output).

NaiveNASlib.SizeAlignFailError — Type

SizeAlignFailError <: Exception

Sizes could not be aligned.

NaiveNASlib.neighbours — Function

neighbours(d::Direction, v)

Return vertices connected to v in direction d.

NaiveNASlib.ΔSizeFailError — Type

ΔSizeFailError <: Exception

Size change could not be solved.

NaiveNASlib.ninconstraint! — Function

ninconstraint!(case, s, v, data)

Add input size constraints for AbstractVertex v using strategy s.

Extra info like the model and variables is provided in data.

NaiveNASlib.newsizes — Function

newsizes(s::AbstractΔSizeStrategy, vertices::AbstractArray{<:AbstractVertex})

Return a vector of new outputs sizes for and a Dict of new input sizes for all provided vertices using the strategy s.

Result vector is index aligned with vertices. Result Dict has a vector of input sizes for each element of vertices which has an input (i.e everything except input vertices).

NaiveNASlib.Input — Type

Input

Represents the input direction, i.e coming from the output of another vertex.

NaiveNASlib.L1NormLinear — Type

L1NormLinear
L1NormLinear()

Add a set of linear constraints to a model to map an expression to a variable which is the L1 norm of that expression.

NaiveNASlib.SumNorm — Type

SumNorm{N<:JuMPNorm} <: JuMPNorm

Sum of ns.

NaiveNASlib.sizeconstraint! — Function

sizeconstraint!(s::AbstractJuMPΔSizeStrategy, v, data)

Add size constraints for AbstractVertex v using strategy s.

Extra info like the model and variables is provided in data.

NaiveNASlib.MaxNormLinear — Type

MaxNormLinear
MaxNormLinear()

Add a set of linear constraints to a model to map an expression to a variable which is the max norm of that expression.

NaiveNASlib.AlignNinToNoutVertices — Type

AlignNinToNoutVertices{V1,V2,S,F} <: AbstractJuMPΔSizeStrategy
AlignNinToNoutVertices(vin::V1, vout::V2, inds; vstrat::S=AlignNinToNout(), fallback::F=ΔSizeFailNoOp())
AlignNinToNoutVertices(vin::V1, vout::V2, inds, vstrat::S, fallback::F)

Same as AlignNinToNout with an additional constraint that nin(s.vin)[s.ininds] == nout(s.vout) where s is a AlignNinToNoutVertices.

Useful in the context of adding edges to vertices to align sizes before the edge has been added.

If it fails, the operation will be retried with the fallback strategy (default ΔSizeFailNoOp).

NaiveNASlib.vertexconstraints! — Function

vertexconstraints!(s::AbstractJuMPΔSizeStrategy, v, data)

Add constraints for AbstractVertex v using strategy s.

Extra info like the model and variables is provided in data.

NaiveNASlib.opposite — Function

opposite(d::Direction)

Return the opposite direction of d.

NaiveNASlib.ScalarSize — Type

ScalarSize

Treat vertices as having a scalar size when formulating the size change problem.

NaiveNASlib.Output — Type

Output

Represents the output direction, i.e coming from the input of another vertex.

NaiveNASlib.sizemodel — Function

sizemodel(s::AbstractJuMPΔSizeStrategy, vertices)

Return a JuMP.Model for executing strategy s on vertices.

NaiveNASlib.ScaleNorm — Type

ScaleNorm{S<:Real,N} <: JuMPNorm
ScaleNorm(scale, n)

Scales result from n with a factor scale.

NaiveNASlib.norm! — Function

norm!(s::L1NormLinear, model, X)

Add a set of linear constraints to a model to map X to an expression X′ which is the L1 norm of X.

Note that it only works for the objective function and only for minimization.

norm!(s::L1NormLinear, model, X)

Add a set of linear constraints to a model to map X to a variable X′ which is the max norm of X.

Note that it only works for the objective function and only for minimization.

NaiveNASlib.accept — Function

accept(case, ::AbstractJuMPΔSizeStrategy, model::JuMP.Model)

Return true of the solution for model is accepted using strategy s.

NaiveNASlib.NeuronIndices — Type

NeuronIndices

Treat vertices as having parameters which represents neurons when formulating the size change problem.

This means that individual indices will be aligned so that the function to the largest extent possible is the same after resizing.

NaiveNASlib.ΔNoutMix — Type

ΔNoutMix{VE, VR, F} <: AbstractJuMPΔSizeStrategy
ΔNoutMix(exact::ΔNout{Exact, VE, F}, relax::ΔNout{Relaxed, VR, F}, fallback::F)

Strategy for changing nout of vertices.

Applies both exact and relax, thereby allowing for having both hard and relaxed size constraints for different vertices.

Can be conveniently created by ΔNout or ΔNin.

NaiveNASlib.SizeCycleDetector.isinsizecycle — Function

isinsizecycle(v)

Return true if removing v leads to a size cycle.

A size cycle is when at least one vertex vi has the impossible size constraint that nout(vi) == nout(vi) + nout(vj) where vj is a vertex in the graph (possibly vi).

This tends to happen when concatenation is followed by an elementwise operation and removing one input vi to the concatenation and replacing it with another vertex vj which is also input to the elementwise operation (possibly through any number of size transparent operations). Since the sum of input sizes to the concatenation must be equal to the input size of the elementwise operation and one out of several inputs to the concatenation must also have the same output size as the input to the elementwise operation we are stuck with an impossible constraint.

As the algorithm is a bit messy (which imo is a sign that it is incorrect), here is a rundown of the idea:

From the queired vertex, trace the graph forwards (output direction) until we hit a non-SizeTransparent vertex. If we encounter a SizeStack vertex store

this vertex in a SeenSizeStack and use it as the current state. If we encounter a SizeInvariant when we have SeenSizeStack as the state, we will store the vertex and the SeenSizeStack as the output. The result from this is thus the set of potentially problematic vertices. If there are no potentially problematic vertices there can not be a size cycle and we return false.

As the posed question is "If I remove v and connect its inputs to its outputs, will there be a size cycle?" we determine if shortcutting v creates a size transparent

path between any input ancestor to v and any of the problematic vertices. To avoid having to check all ancestors, we just examine the set which terminates any size transparency using findterminating, knowing that if v and a problematic vertex share such a terminating ancestor, then shortcutting v creates a size transparent path to the problematic vertex which may not have been there before. If there are no problematic vertices which share a common terminating ancestor with v we return false.

For each problematic vertex vp and associated set of terminating ancestors VPa which passes the check in 2 we now check if any of the vertices in VPa sees vp as

problematic for the same reason as v does, meaning that they both pass through the same concatenation at least one time before hitting vp. If not, it basically means that v is alone in a noop concatenation and there is no vj from the first paragraph to create an impossible constraint. If this applies to all problematic vertices we don't have a size cycle and return false.