Extra Vertex Traits

Imported to namespace by

using NaiveNASlib.Advanced

NaiveNASlib.NamedTrait — Type

NamedTrait <: DecoratingTrait
NamedTrait(name, base)

Trait which attaches name to a vertex. Calling name(v) on a vertex with this trait returns name.

NaiveNASlib.AfterΔSizeTrait — Type

AfterΔSizeTrait <: DecoratingTrait
AfterΔSizeTrait(strategy::S, base::T)

Calls after_Δnin(strategy, v, Δs, ischanged) and after_Δnout(strategy, v, Δ, ischanged) after a size change for the vertex v which this trait is attached to.

NaiveNASlib.named — Function

named(name)

Return a function t -> NamedTrait(name, t) intended to reduce a bit of the verbosity when using traitdecoration.

Intended to be composable with other similar functions through ∘.

NaiveNASlib.logged — Function

logged(args...;kwargs...)

Return a function t -> AfterΔSizeTrait(logafterΔsize(args...;kwargs...) indended to reduce a bit of the verbosity when using traitdecoration.

Intended to be composable with other similar functions through ∘.

NaiveNASlib.validated — Function

validated(args...;kwargs...)

Return a function t -> AfterΔSizeTrait(validateafterΔsize(args...;kwargs...), t) indended to reduce a bit of the verbosity when using traitdecoration.

Intended to be composable with other similar functions through ∘.