Graph Operations

NaiveNASlib.CompGraph — Type

CompGraph
CompGraph(input::AbstractVertex, output::AbstractVertex)
CompGraph(input::AbstractVector{<:AbstractVertex}, output::AbstractVertex)
CompGraph(input::AbstractVertex, output::AbstractVector{<:AbstractVertex})

Basic graph for computation. While not strictly neccessary to compute anything, it makes it easier to keep track of things.

Examples

julia> using NaiveNASlib

julia> v1 = inputvertex("in1", 1) + inputvertex("in2", 1);

julia> v2 = invariantvertex(x -> 3x, v1);

julia> CompGraph(inputs(v1), v2)(2,3) # (2 + 3) * 3
15

julia> CompGraph(inputs(v1), [v1, v2])(2,3)
(5, 15)

NaiveNASlib.inputs — Method

inputs(g::CompGraph)

Return the inputs vertices of g.

NaiveNASlib.outputs — Method

outputs(g::CompGraph)

Return the output vertices of g.

NaiveNASlib.vertices — Function

vertices(g::CompGraph)

Return an topologically sorted array of all vertices in the graph g.

Examples

julia> ins = InputVertex.(1:2);

julia> v1 = CompVertex(+, ins...);

julia> v2 = CompVertex(*, v1, ins[2]);

julia> graph = CompGraph(ins, v2);

julia> vertices(graph)
4-element Array{AbstractVertex,1}:
 InputVertex(1)
 InputVertex(2)
 CompVertex(+), inputs=[InputVertex(1), InputVertex(2)]
 CompVertex(*), inputs=[CompVertex(+), InputVertex(2)]

NaiveNASlib.nvertices — Function

nvertices(g::CompGraph)

Return the number of vertices in the graph.

NaiveNASlib.findvertices — Function

findvertices(g::CompGraph, vname::AbstractString)

Return all vertices for which name(v) == vname`.

findvertices(g::CompGraph, vname::Regex)

Return all vertices for which vpat matches name(v).